file fixes, move code from core -> roi
